% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hList.R
\docType{methods}
\name{hList}
\alias{hList}
\title{Extract checksum value}
\usage{
hList(fList, dir, quick = TRUE, csalgorithm = "xxhash64")
}
\arguments{
\item{fList}{A vector listing object from which we want to extract chekcsum value.}

\item{dir}{A character string representing path to file.}

\item{quick}{A logical argument. If TRUE, checksum is compiled using the combination of the filename and its size.
If FALSE, cheksum is compiled using the object. Default is FALSE.}

\item{csalgorithm}{A character string representing the algorithms used in the digest function. Default is "xxhash64".}
}
\value{
The hList function returns a dataframe where filename, checksum value from the object (checksumFile), checksum
        value from the combination of filename and file size and the algorithm used to compute checksum values are stored.
}
\description{
The function hList extract checksum value from a list of file and return the value its respective filename as an R
object. The function is for now using the algorithm 'xxhash64'.Checksum is compiled using the digest function.
}
\examples{
outdir<- tempdir()
file.list<- list.files(outdir)
hfile <-hList(file.list, dir = outdir, quick = TRUE)
}
\author{
Melina Houle
}
