% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{methods}
\name{.hdf}
\alias{.hdf}
\title{Template data.frame to use for file hash table}
\usage{
.hdf()
}
\description{
Internal function. Returns an empty \code{data.frame} with columns:
\code{Filename}, \code{checksumFile}, \code{checksumSize}, and \code{algorithm}.
}
\author{
Alex Chubaty
}
\keyword{internal}
