% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logHash.R
\docType{methods}
\name{logHash}
\alias{logHash}
\title{logHash write its required argument \code{hashdata} to a SQLite database.}
\usage{
logHash(hashdata, dbHash = "dbHash.sqlite")
}
\arguments{
\item{hashdata}{A \code{data.frame} containing \code{filename}, \code{checksumFile},
\code{checksumSize} and \code{algorithm} for each file are stored.}

\item{dbHash}{A character string indicating the name of the database where checksum values are stored.
If the database does not exist, the function will create it and checksum will be computed.
Default is \code{"dbHash.sqlite"}
.}
}
\value{
Invoked for its side-effect of logging filenames and checksum values into the SQL database.
}
\description{
If the SQLite database already exists, \code{hashdata} argument is append if not already present.
}
\examples{
destfile <-tempdir()
dbHash <- file.path(destfile, "dbHash.sqlite")
file.list<- list.files(destfile)
hfile <- hList(file.list, destfile, quick = TRUE)
logHash(hfile, dbHash)

}
\author{
Melina Houle
}
