% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listWebData.R
\name{webDatabases}
\alias{webDatabases}
\title{Grab the current version of the urls from the git repository.}
\usage{

  webDatabases(dbUrl = "https://raw.githubusercontent.com/PredictiveEcology/webDatabases/master/R/webDatabases.R",
  local = FALSE, wide = FALSE)
}
\arguments{
\item{dbUrl}{Character string where to look for web database. Defaults to:
\code{https://raw.githubusercontent.com/PredictiveEcology/webDatabases/master/R/webDatabases.R}}

\item{wide}{Logical. If \code{TRUE}, the result is returned in long format.}

\item{wide}{Logical. If \code{TRUE}, returns the wide form of database. Default \code{FALSE}}
}
\value{
Data.table with 4 \code{columns}, \code{dataset}, \code{url}, \code{password},
and \code{files}, keyed by \code{dataset},
 \code{files}
}
\description{
Running this will get latest version of the urls, returned as a data.table.
}
