% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashDownload.R
\docType{methods}
\name{hashDownload}
\alias{hashDownload}
\title{Download and store hash value from downloaded file}
\usage{
hashDownload(urls, destfile, checkhash = TRUE, quick = FALSE,
  dbHash = "dbHash.sqlite", cascade = FALSE, quiet = TRUE)
}
\arguments{
\item{urls}{A character string. Represents the url of file to be downloaded.}

\item{destfile}{A character string. Represents the name where resulting file is saved.}

\item{checkhash}{Logical. If \code{TRUE} (default), check if file exists locally
and cross-check the file's checksum value with that logged
from a previous event, skippng redownload if checksum matches.
When checksums don't match or file doesn't exist locally,
the file is downloaded and its checksum computed.
If \code{FALSE}, the function is executed even if the file
is found locally.}

\item{quick}{Logical. If \code{TRUE}, checksum is compiled using the combination
of the filename and its size.
If \code{FALSE} (default), cheksum is compiled using the object.}

\item{dbHash}{Character string. The path to the database file where checksum
value of file is logged.
If the database does not yet exist, one is created. Default is \code{"dbHash.sqlite"}.}

\item{cascade}{Logical. If \code{FALSE}, file is untar and/or unzip. Default is \code{FALSE}.}

\item{quiet}{Logical. If \code{TRUE}, suppress status messages (if any), and the progress bar.}
}
\value{
Invoked for its side-effect of downloading files to the \code{destfile/} directory.
}
\description{
To avoid redownloading existing files, verify if the file exists locally.
If the file exists, compare the checksum value to that logged by original download.
}
\examples{
u <-"http://ftp.geogratis.gc.ca/pub/nrcan_rncan/archive/vector/cli_itc_50k/land_use/L040J03.zip"
hashDownload(urls = u, destfile = tempdir(), checkhash = FALSE, cascade = FALSE)
}
\author{
Melina Houle
}
\keyword{internal}
