% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashUntar.R
\docType{methods}
\name{hashUntar}
\alias{hashUntar}
\title{Extract files from a tar archive and log checksum value to avoid repeating event}
\usage{
hashUntar(tarfile, destfile, checkhash = TRUE, quick = FALSE,
  dbHash = "dbHash.sqlite")
}
\arguments{
\item{tarfile}{A character string. Represents path to tarfile.}

\item{destfile}{A character string. Contains the name where untar file is saved.}

\item{checkhash}{Logical. If \code{TRUE} (default), check if untarred file exists locally and compare
checksum value with checksum logged from previous event. When checksums don't
match or ubtar file doesn't exist , \code{untar} and \code{digest} are performed.
If \code{FALSE}, only \code{untar} is performed.}

\item{quick}{Logical. If \code{TRUE}, \code{digest} is performed using the combination of
untarred filename and its size. If \code{FALSE} (default), \code{digest} is performed
using the object.}

\item{dbHash}{A character string. Represents path to SQLite database file where checksum value
from \code{digest} is logged. If the database doesn't exist, one is created. Default is
\code{"dbHash.sqlite"}.}
}
\value{
Untarred/unzipped \code{tarfile} in a subfolder under \code{destfile} using
        \code{basename{tarfile}} name.
}
\description{
Prior to untar, the function check if untar was previously performed (untarred file exist
locally) and compare checksum value from previous event when present. Untar is performed
when file doesn't exists or checksums don't match.
}
\examples{
urlpath <- "ftp://sidads.colorado.edu/pub/DATASETS/NOAA/G02171/Hudson_Bay/2006"
urlfile <-  "cis_SGRDRHB_20060904_pl_a.tar"
url <- file.path(urlpath, urlfile)
hashDownload(url, destfile = tempdir(), cascade = FALSE)
tar<- file.path(tempdir(), basename(url))
hashUntar(tar, tempdir(), checkhash= FALSE)
}

}
\author{
Melina Houle
}
