% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashUnzip.R
\docType{methods}
\name{hashUnzip}
\alias{hashUnzip}
\title{Extract files from a zip archive downloaded and log checksum value}
\usage{
hashUnzip(zipfile, destfile, checkhash = TRUE, quick = FALSE,
  dbHash = "dbHash.sqlite")
}
\arguments{
\item{zipfile}{A character vector. Contains path to zipfile.}

\item{destfile}{A character string. Represents the name where unzip file is saved.}

\item{checkhash}{Logical. If \code{TRUE}, check if unzip file is found locally
and cross-check its checksum value with value logged in dbHash
from previous unzip event.
When checksums match, a message raises indicating the file is
already properly unzip.
When checksums doesn't match or unzip file doesn't exist locally,
unzip occurs and checksum computed.
If \code{FALSE} (default), unzip occurs even if the files exist locally.}

\item{quick}{Logical. If \code{TRUE}, checksum is coputed using the combination
of the filename and its size.
If \code{FALSE} (default), checksum is computed using the object.}

\item{dbHash}{A character string. The path to the database file where file
checksum value is logged.
If the named database does not yet exist, one is created.
Default is \code{"dbHash.sqlite"}.}
}
\value{
Invoked for its side-effect of saving unzipped file in a subdirectory
        using the zipfile basename in the \file{destfile/} directory.
}
\description{
unzip dataset previously downloaded from url, first checking if
unzip was previously performed (unzip file exists locally).
}
\examples{
url<-"http://ftp.geogratis.gc.ca/pub/nrcan_rncan/archive/vector/cli_itc_50k/land_use/L040J03.zip"
hashDownload(url, destfile = tempdir(), checkhash= FALSE, cascade = FALSE)
zip <- file.path(tempdir(), basename(url))
hashUnzip(zip, tempdir(), checkhash= FALSE)

}
\author{
Melina Houle
}
