% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hList.R
\docType{methods}
\name{hList}
\alias{hList}
\title{Extract checksum value from file}
\usage{
hList(fList, destfile, quick = FALSE, csalgorithm = "xxhash64")
}
\arguments{
\item{fList}{A character string. Represents name of file.}

\item{destfile}{A character string. Represents the path name to file.}

\item{quick}{Logical. If \code{TRUE}, \code{digest} is performed using the combination of
unzipped filename and its size. If \code{FALSE} (default), \code{digest} is performed
using the object.}

\item{csalgorithm}{A character string.  The algorithm to be used by \code{digest} (see Details of
\code{\link[digest]{digest}}). Default is "xxhash64".}
}
\value{
A \code{dataframe} containing \code{filename}, \code{checksumFile}, \code{checksumSize}
       and \code{algorithm} for each file are stored
}
\description{
The function stores in a dataframe the filename, checksum file value, checksum size value and
the algorithm used to perform checksum. Checksum is compiled using the digest function (see
Details of \code{\link[digest]{digest}}).
}
\examples{
outdir<- tempdir()
file.list<- list.files(outdir)
hfile <- hList(file.list, destfile = outdir, quick = TRUE)

}
\author{
Melina Houle
}
