% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hList.R
\docType{methods}
\name{hList}
\alias{hList}
\title{Extract checksum value from file}
\usage{
hList(fList, destfile, quick = TRUE, csalgorithm = "xxhash64")
}
\arguments{
\item{fList}{A character vector representing filename}

\item{destfile}{A character string giving file to path to file.}

\item{quick}{A logical argument. If TRUE, checksum is compiled using the combination
of the filename and its size. If FALSE, cheksum is compiled using the object.
Default is FALSE.}

\item{csalgorithm}{A character string representing the algorithms used by the digest
function. Default is "xxhash64".}
}
\value{
The hList function returns a dataframe where filename, checksum value from the
        object (checksumFile), checksum value from the combination of filename and file
        size and the algorithm used to compute checksum values are stored.
}
\description{
The function stores in a dataframe the filename and its checksum value. Checksum is compiled
using the digest function (see Details of \code{\link[digest]{digest}}).
}
\examples{
outdir<- tempdir()
file.list<- list.files(outdir)
hfile <- hList(file.list, destfile = outdir, quick = TRUE)

}
\author{
Melina Houle
}
