% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hList.R
\docType{methods}
\name{hList}
\alias{hList}
\title{Extract checksum value from file}
\usage{
hList(fList, destfile, quick = FALSE, csalgorithm = "xxhash64")
}
\arguments{
\item{fList}{A character vector representing filename.}

\item{destfile}{A character string giving file to path to file.}

\item{quick}{Logical. If \code{TRUE}, checksum is compiled using the combination
of the filename and its size.
If \code{FALSE} (default), cheksum is compiled using the object.}

\item{csalgorithm}{A character string representing the algorithms used by the
digest function. Default is \code{"xxhash64"}.}
}
\value{
\code{data.frame} where filename, checksum value from the object
        (checksumFile), checksum value from the combination of filename and file
        size and the algorithm used to compute checksum values are stored.
}
\description{
Stores, in a \code{data.frame}, the filename and its checksum value (computed
using the \code{\link[digest]{digest}}).
}
\examples{
  outdir <- tempdir()
  file.list <- list.files(outdir)
  webDatabases::hList(file.list, destfile = outdir, quick = TRUE)

}
\author{
Melina Houle
}
\keyword{internal}
