% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listWebData.R
\docType{methods}
\name{listWebData}
\alias{listWebData}
\title{Retrive file available to download.}
\usage{
listWebData(urlTble, datasetName)
}
\arguments{
\item{urlTble}{A \code{data.table} containing available \code{url}, \code{datasetName},
\code{password} and code{filenames} found using the \code{url}. \code{urltble}
is provided within the package as \code{urls} object.}

\item{datasetName}{A character string. Name of the dataset from which \code{url} and
\code{password} can be derived from the \code{urltble}.}
}
\value{
Vector of available url to download
}
\description{
The function produce a character vector of file available to download. It uses a preset URL
table in which a list of relevant dataset name, their associate url and password are stored.
To retrive available file, the function derive URL and password using the datasetname.
}
\examples{
library(data.table)
dt<- data.table(dataset = c("NFDB"),
               url = c("http://cwfis.cfs.nrcan.gc.ca/downloads/nfdb/fire_poly/current_version/"),
               password= c(NA))
listWebData(dt, datasetName = "NFDB")
}
\author{
Melina Houle
}
