% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logHash.R
\docType{methods}
\name{logHash}
\alias{logHash}
\title{Log checksum from downloaded files into a SQLite database}
\usage{
logHash(hashdata, dbHash = "dbHash.sqlite")
}
\arguments{
\item{hashdata}{A \code{data.frame} containing \code{filename}, \code{checksumFile},
\code{checksumSize} and \code{algorithm} for each file are stored.}

\item{dbHash}{A character string indicating the name of the database where checksum values are stored.
If the database does not exist, the function will create it and checksum will be computed.
Default is \code{"dbHash.sqlite"}
.}
}
\value{
Invoked for its side-effect of logging filenames and checksum values into the SQL database.
}
\description{
Log compiled checksum values from a specific file into a predetermine dbHash
sqlite database following SQL language requirements.
}
\examples{
destfile <-tempdir()
dbHash <- file.path(destfile, "dbHash.sqlite")
file.list<- list.files(destfile)
hfile <- hList(file.list, destfile, quick = TRUE)
logHash(hfile, dbHash)

}
\author{
Melina Houle
}
