% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logHash.R
\docType{methods}
\name{logHash}
\alias{logHash}
\title{Write \code{hashdata} to SQLite database}
\usage{
logHash(hashdata, dbHash = "dbHash.sqlite")
}
\arguments{
\item{hashdata}{A \code{data.frame} containing \code{filename}, \code{checksumFile},
\code{checksumSize} and \code{algorithm} for each file are stored.}

\item{dbHash}{Character string. The path to the database file where checksum
value of file is logged.
If the database does not yet exist, one is created. Default is \code{"dbHash.sqlite"}.}
}
\value{
Invoked for its side-effect of logging filenames and checksum values into the SQL database.
}
\description{
If the SQLite database already exists, \code{hashdata} argument is append if not already present.
}
\examples{
\dontrun{
destfile <-tempdir()
dbHash <- file.path(destfile, "dbHash.sqlite")
file.list<- list.files(destfile)
hfile <- hList(file.list, destfile, quick = TRUE)
logHash(hfile, dbHash)
}
}
\author{
Melina Houle
}
\keyword{internal}
