% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listWebData.R
\name{webDatabases}
\alias{webDatabases}
\title{Grab the current version of the urls from the git repository.}
\usage{
webDatabases(wide = TRUE)
}
\arguments{
\item{wide}{Logical. If TRUE, the result is returned in long format.}
}
\value{
Data.table with 4 columns, dataset, url, password (currently all NA), files, keyed by dataset, files
}
\description{
Running this will get latest version of the urls, returned as a data.table.
}
