% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listWebData.R
\docType{methods}
\name{urls}
\alias{urls}
\title{Table of relevant dataset accessible using url}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 8 rows and 4 columns.}
\usage{
urls
}
\value{
\code{data.table} containing dataset available for download.
}
\description{
Stores dataset name with their respective url and username/password.
The table is used by \code{listWebDatabases} to retrive file available to
download by using only the dataset name.
}
\author{
Melina Houle
}
\keyword{datasets}
