% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listWebData.R
\docType{methods}
\name{urls}
\alias{urls}
\title{Table of relevant dataset accessible using url}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 8 rows and 4 columns.}
\usage{
urls
}
\value{
data.table object containing dataset available for download.
}
\description{
An R object that stores dataset name with their respective url and username/password. The table is used
by the function listWebDatabases to retrive file available to download by using only the dataset name.
}
\examples{
library(data.table)
dt<- data.table(
  dataset = c("NFDB"),
  url = "http://cwfis.cfs.nrcan.gc.ca/downloads/nfdb/fire_poly/current_version/",
  password = NA,
  files = "NFDB_poly.zip"
)

}
\author{
Melina Houle
}
\keyword{datasets}
